requirements.txt
# Computing
numpy==1.17.4


# Visualization & I/O
matplotlib==3.1.1
pillow==6.2.1
ipython==7.9.0
open3d
